Initial commit. Features a CardSwitcherCard for multiple entries.
CardSwitcherCard allows the current card to switch when Mozart is sent a
message with parameters. The switcher needs to always be the last card
on the card stack when passfocussing the application. The other cards
are in the cache and the correct one is loaded by the switcher when the
fous is back with Mozart.

This allows us to be playing music and for the internet radio
application to take over directly, without displaying old music playing
metadata.